+++ lib/src/anim_type.dart
@@ -1,0 +1,11 @@
+enum FadeType {
+  None,
+  FadeIn,
+  FadeOut,
+}
+
+enum TranslateType {
+  None,
+  FromMyPosToNewPos,
+  FromNewPosToMyPos,
+}